requirements.txt
openai>=1.0.0
googlemaps
tqdm
pandas
python-dotenv
httpx
tiktoken
pyyaml
